% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.extras.R
\name{compute_clustered_standard_error.gmnl}
\alias{compute_clustered_standard_error.gmnl}
\title{Compute Clustered Standard Errors for a gmnl Model}
\usage{
compute_clustered_standard_error.gmnl(model)
}
\arguments{
\item{model}{A `gmnl` model object for which the clustered standard errors will be computed.}
}
\value{
Prints the clustered standard errors covariance matrix.
}
\description{
Computes the clustered standard errors for a gmnl model object
using the sandwich `vcovCL()` function.
}
