% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.extras.R
\name{compute_elasticities_cross.gmnl}
\alias{compute_elasticities_cross.gmnl}
\title{Compute elasticities for a given generalized multinomial logit (gmnl) model}
\usage{
compute_elasticities_cross.gmnl(gmnl_model, wrta)
}
\arguments{
\item{gmnl_model}{A gmnl model object.}

\item{wrta}{A string indicating the alternative to compute elasticities for.}
}
\value{
A gmnl model object with updated elasticities and predicted probabilities.
}
\description{
This function computes the cross elasticities for a given gmnl model by changing the value of the cost variable for a specific alternative and then computing the change in predicted probabilities.
}
