% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.extras.R
\name{compute_elasticities_direct.gmnl}
\alias{compute_elasticities_direct.gmnl}
\title{Compute elasticities for a GMNL model}
\usage{
compute_elasticities_direct.gmnl(gmnl_model, wrta)
}
\arguments{
\item{gmnl_model}{A GMNL model object}

\item{wrta}{A character string specifying the reference alternative with
respect to which the elasticities are to be computed.}
}
\value{
A gmnl object with the updated elasticities, along with the data summary
}
\description{
Given a GMNL model and a reference alternative, this function computes the
elasticity of each attribute with respect to the reference alternative. The
function returns a summary of the elasticities by mode.
}
