% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.extras.R
\name{DeltaCI}
\alias{DeltaCI}
\title{Computes Variety of Confidence Interval Data}
\usage{
DeltaCI(res, Q, kcoe, costcoe, level, qw, tpw)
}
\arguments{
\item{res}{an object of class glmn}

\item{Q}{number of random draws from standard distributions (normal or uniform)}

\item{kcoe}{vector of positions (in the res$coe vector) of the parameters for which wtp is to be computed}

\item{costcoe}{position (in the res$coe vector) of the cost parameter}

\item{level}{vector of confidence level to compute confidence and prediction intervals (i.e. 0.9, 0.95…)}

\item{qw}{value (the same for all parameter for which wtp is to be computed) at which upper tail probability
for wtp distribution is to be computed}

\item{tpw}{probability (the same for all parameters for which wtp is to be computed) at which quantiles
leaving that probability in the upper tail of wtp distribution are to be computed}
}
\value{
vector of : "WTP", "se.mean.WTP","pred.err.WTP","CI.mean.WTP","PI.WTP","quant.WTP","tail.prob.WTP"
WTP : Willingness to pay point estimate
se.mean.WTP : Willingness to pay point estimate
pred.err.WTP : Standard error for the mean WTP
CI.mean.WTP : Prediction error for WTP
PI.WTP : Prediction interval for WTP
quant.WTP : Quantile such that tpw% of the population has a larger WTP
tail.prob.WTP : Probability of observing a WTP larger than qw
}
\description{
Computes Variety of Confidence Interval Data
}
