% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.extras.R
\name{F_inv}
\alias{F_inv}
\title{Compute Quantiles for Mixture of Normal Distributions}
\usage{
F_inv(p, w, u, s, br = c(-1000, 1000))
}
\arguments{
\item{p}{CDF corresponding to the quantile to be computed}

\item{w}{vector of weights for the normal densities in the mixture}

\item{u}{vector of means for the normal densities in the mixture}

\item{s}{vector of std dev for the normal densities in the mixture}

\item{br}{vector of lower and upper bounds of the the interval in which the quantile is to be searched for}
}
\value{
the quantile corresponding to the cumulative density function p
}
\description{
Compute Quantiles for Mixture of Normal Distributions
}
