% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.extras.R
\name{Fa}
\alias{Fa}
\title{Cumulative Density Function for Mixture of Normal Distributions}
\usage{
Fa(x, w, u, s)
}
\arguments{
\item{x}{value at which CDF needs to be computed}

\item{w}{vector of weights for the normal densities in the mixture}

\item{u}{vector of means for the normal densities in the mixture}

\item{s}{vector of std dev for the normal densities in the mixture}
}
\value{
cumulative density function of x
}
\description{
Cumulative Density Function for Mixture of Normal Distributions
}
