% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.methods.R
\name{getSummary.gmnl}
\alias{getSummary.gmnl}
\title{Get Model Summaries for Use with "mtable"}
\usage{
getSummary.gmnl(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{a \code{gmnl} object,}

\item{alpha}{level of the confidence intervals,}

\item{...}{further arguments,}
}
\description{
A generic function to collect coefficients and summary statistics from a \code{gmnl} object.
}
\details{
For more details see package \pkg{memisc}
}
