% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.data.R
\name{gmnl.data}
\alias{gmnl.data}
\title{GMNL Data}
\usage{
gmnl.data(
  data,
  choice = NULL,
  shape = c("long", "wide"),
  varying = NULL,
  sep = ".",
  alt.var = NULL,
  chid.var = NULL,
  alt.levels = NULL,
  id.var = NULL,
  group.var = NULL,
  opposite = NULL,
  drop.index = FALSE,
  ranked = FALSE,
  subset = NULL,
  ...
)
}
\arguments{
\item{data}{a `data.frame`,}

\item{choice}{the variable indicating the choice made: it can be
either a logical vector, a numerical vector with 0 where the
alternative is not chosen, a factor with level 'yes' when the
alternative is chosen}

\item{shape}{the shape of the `data.frame`: whether `long` if each
row is an alternative or `wide` if each row is an observation}

\item{varying}{the indexes of the variables that are alternative
specific, sometimes seen as xxxx.1, xxxx.2, etc}

\item{sep}{the charachter  that separates the variable name
and the alternative name (only for `wide` data)}

\item{alt.var}{the name of the variable that contains the
alternative index (for a `long` `data.frame` only) or the name
under which the alternative index will be stored (the default
name is `alt`),}

\item{chid.var}{the name of the variable that contains the choice
index or the name under which the choice index will be stored,}

\item{alt.levels}{the name of the alternatives: if null, for a
`wide` data.frame, they are guessed from the variable names and
the choice variable (both should be the same), for a `long`
`data.frame`, they are guessed from the `alt.var` argument,}

\item{id.var}{the name of the variable that contains the individual
index if any,}

\item{group.var}{the name of the variable that contains the group
index if any,}

\item{opposite}{returns the opposite of the specified variables,}

\item{drop.index}{should the index variables be dropped from the
`data.frame`}

\item{ranked}{a logical value which is true if the response is a ranked variable}

\item{subset}{logical expression which defines the subset of
observations to be selected}

\item{...}{more arguments to specify reshaping of the data}
}
\value{
gmnl.data object that inherits class dfidx
}
\description{
GMNL Data
}
