% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.draws.R
\name{halton}
\alias{halton}
\title{The Halton sequence helps us generate quasi random numbers in maximum likelihood
problems to reduce computation time. If 'halton' given to make.draws() is NA,
the default value of the primes in vector 'prime' that are used in order, as well
as a default value are dropped. If 'halton' is provided in the gmnl() call, it
should be a list with elements that will go into haltons$prime (primes used) and
elements for haltons$drop (number of elements dropped).}
\usage{
halton(prime = 3, length = 100, drop = 10)
}
\arguments{
\item{prime}{prime to generate the Halton draws on - default value of 3}

\item{length}{how many draws to generate - default value of 100}

\item{drop}{first n samples to drop to remove correlation}
}
\description{
Halton function from mlogit package
}
