% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.extras.R
\name{KRCI}
\alias{KRCI}
\title{Krinsky & Robb}
\usage{
KRCI(res, Q, B, kcoe, costcoe, level, qw, tpw)
}
\arguments{
\item{res}{an object of class gmnl}

\item{Q}{number of random draws from standard distributions (normal or uniform)}

\item{B}{number of bootstrap samples}

\item{kcoe}{vector of positions (in the res$coe vector) of the parameters for which wtp is to be computed}

\item{costcoe}{position (in the res$coe vector) of the cost parameter}

\item{level}{vector of confidence level to compute confidence and prediction intervals (i.e. 0.9, 0.95…)}

\item{qw}{value (the same for all parameter for which wtp is to be computed) at which upper tail probability
for wtp distribution is to be computed}

\item{tpw}{probability (the same for all parameters for which wtp is to be computed) at which quantiles leaving
that probability in the upper tail of wtp distribution are to be computed}
}
\value{
vector of : "se.mean.WTP","pred.err.WTP","CI.mean.WTP","PI.WTP","quant.WTP","se.quant.WTP","CI.quant.WTP","tail.prob.WTP","se.tail.prob.WTP","CI.tail.prob.WTP"
}
\description{
Krinsky & Robb
}
