% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.methods.R
\name{mcfadden_adj.gmnl}
\alias{mcfadden_adj.gmnl}
\title{McFadden R^2 Adjusted}
\usage{
mcfadden_adj.gmnl(object, returnObj = FALSE)
}
\arguments{
\item{object}{gmnl object}

\item{returnObj}{optional TRUE/FALSE for whether a gmnl object is returned containing the mcfadden value}
}
\value{
mcfadden value OR gmnl object containing mcfadden value
}
\description{
McFadden R^2 Adjusted
}
