% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.methods.R
\name{rho2.gmnl}
\alias{rho2.gmnl}
\title{Rho^2 Value}
\usage{
rho2.gmnl(object, returnObj = FALSE)
}
\arguments{
\item{object}{a gmnl object}

\item{returnObj}{optional bool, TRUE to return gmnl object with Rho^2 appended
FALSE to just return Rho^2 value}
}
\value{
gmnl object or Rho^2 value
}
\description{
Rho^2 Value
}
