% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.methods.R
\name{rho2_adj.gmnl}
\alias{rho2_adj.gmnl}
\title{Adjusted Rho^2 Value}
\usage{
rho2_adj.gmnl(object, returnObj = FALSE)
}
\arguments{
\item{object}{gmnl object of a computed model}

\item{returnObj}{optional bool, TRUE to return gmnl object with adj Rho^2 appended
FALSE to just return Rho^2 value}
}
\value{
gmnl object or adjusted Rho^2 value
}
\description{
Adjusted Rho^2 Value
}
