% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.extras.R
\name{summary.KRCI}
\alias{summary.KRCI}
\title{Summary function for KRCI}
\usage{
\method{summary}{KRCI}(object, ...)
}
\arguments{
\item{object}{KRCI object}

\item{...}{other}
}
\description{
Summary function for KRCI
}
